Print a warning if the kernel is allowed to move threads about
