+++ .github/workflows/ci-julia-nightly.yml
@@ -22,22 +22,30 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        version:
-          - '1.5'
-          - '1'
-          - 'nightly'
+        arch:
+          - x64
+          - x86
+        exclusive:
+          - 'false'
         os:
           - ubuntu-latest
           - macOS-latest
           - windows-latest
         threads:
           - '5'
-        arch:
-          - x64
-          - x86
+        version:
+          - '1.5'
+          - '1'
+          - 'nightly'
         exclude:
           - os: macOS-latest
             arch: x86 # 32-bit Julia binaries are not available on macOS
+        include:
+          - exclusive: 'true'
+            threads: '2'
+            arch: x64
+            os: ubuntu-latest
+            version: '1'
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
@@ -57,6 +65,7 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
         env:
+          JULIA_EXCLUSIVE: ${{ matrix.exclusive }}
           JULIA_NUM_THREADS: ${{ matrix.threads }}
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
