+++ test/aqua.jl
@@ -1,0 +1,3 @@
+@testset "Aqua.jl" begin
+    @time Aqua.test_all(ThreadingUtilities)
+end
