+++ test/internals.jl
@@ -1,0 +1,3 @@
+@testset "Internals" begin
+    @test ThreadingUtilities.store!(pointer(UInt[]), (), 1) == 1
+end
